plugins/gmail-api-mailer/requirements.txt
google-api-python-client>=2.169.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
